Fix search sorting
